Removed the rest of the KeyError references
in the Errors and Exceptions lesson.

I chose to keep the KeyError in the first challenge, because it is
a common thing to run into an exception that you've never heard of
before; students should have no problems answering the 6 questions
in the challenge despite not seeing a KeyError before. If anything,
it makes this challenge a bit cleaner, as we're mostly asking for
generic answers rather that should be the same regardless of the
error type.