Style fixes for challenge solutions.
Sometimes it's weird to put the solution in an {.output} cell,
because it renders as fixed width. I think we should have a specfic
way of styling the solutions, but for now, I think it's fine to
just have it inline like it's a question -- perhaps we should
get rid of the questions in the solution set?

Also, where it made sense to keep it a code cell, I made it look
more like code.