Remove reference to size of indent
I think referring to the size of indents at this point could cause confusion. Given that it's not the first time we refer to indents, a reader might think that there is something about functions in particular that means that means they should be indented by four spaces. Also, this is already better dealt with in lesson 7.