Fixed incorrect line numbers in 08-defensive.md
The line numbers referred to in the normalize_rectangle function of defensive programming were incorrect.  My guess is the comment line was added later, and the line numbers were not updated.  I corrected "The preconditions on lines 2, 4, and 5 catch invalid inputs:" to be "The preconditions on lines 3, 5, and 6 catch invalid inputs:" and "...we realize that line 10..." to be "...we realize that line 11..."

Also added the line numbers for the post-conditions "The post-conditions on lines 17 and 18...".  This clarifies to learners which assertions were preconditions and which ones were post-conditions