06-func.md: Remove the word "simple" (/issues/691)
I feel that using the word "simple" can be a bit dismissive. I don't think we can assume that novices would agree that the tests outlined would be simple for them to come up with when they're first starting out with Python. I think this can be reworded to convey a similar meaning without being as dismissive.